test: Build files for Windows CMake (googleapis/google-cloud-cpp-spanner#595)
* Add files for Windows CMake build

* Avoid warning

* Only remove filed when KOKORO_ARTIFACTS_DIR is set

* Add using testing::DoAll

* Avoid secure warnings on Windows

* Install testing library

* Do not build tests for google-cloud-cpp

* Add CMAKE_CXX_FLAGS

* Avoid secure warnings on Windows

* Avoid conversion warning on Windows

* Set values for integration tests

* Avoid conversion warning on Windows

* Add more environment variable for tests

* Add another environment variable for tests

* Run tests serially

* Exclude integration tests

* Use the correct directory

* Fix for clang-tidy

* Change ExtractTupleValues::i to int

* Set _CRT_SECURE_NO_WARNINGS in CMake file

+++ google/cloud/spanner/internal/time.cc
@@ -37,7 +37,7 @@ google::protobuf::Duration ToProto(std::chrono::nanoseconds ns) {
   ns -= std::chrono::duration_cast<std::chrono::nanoseconds>(s);
   google::protobuf::Duration proto;
   proto.set_seconds(s.count());
-  proto.set_nanos(ns.count());
+  proto.set_nanos(static_cast<int>(ns.count()));
   return proto;
 }
 
@@ -145,7 +145,7 @@ google::protobuf::Timestamp ToProto(Timestamp ts) {
   auto ns = std::chrono::duration_cast<std::chrono::nanoseconds>(p.second);
   google::protobuf::Timestamp proto;
   proto.set_seconds(p.first.count());
-  proto.set_nanos(ns.count());
+  proto.set_nanos(static_cast<int>(ns.count()));
   return proto;
 }
